+++ ohsome_quality_api/geodatabase/select_intersection.sql
@@ -4,9 +4,10 @@ WITH bpoly AS (
 )
 SELECT
     -- ratio of area within coverage (empty if outside, between 0-1 if intersection)
-    ST_Area (ST_Intersection (bpoly.geom, coverage.geom)) / ST_Area (bpoly.geom) as area_ratio
+    ST_Area (ST_Intersection (bpoly.geom, coverage.geom)) / ST_Area (bpoly.geom) as area_ratio,
+    ST_AsGeoJSON (ST_Intersection (bpoly.geom, coverage.geom)) AS geom
 FROM
     bpoly,
-    eubucco_v0_1_coverage_simple coverage
+    {table_name} coverage
 WHERE
     ST_Intersects (bpoly.geom, coverage.geom)