+++ tests/integrationtests/indicators/test_building_comparison.py
@@ -8,52 +8,96 @@ import pytest
 
 from ohsome_quality_api.indicators.building_comparison.indicator import (
     BuildingComparison,
-    get_sources,
 )
 from tests.integrationtests.utils import oqapi_vcr
 
 
-@pytest.fixture(scope="class")
+@pytest.fixture
 def mock_get_building_area(class_mocker):
-    async_mock = AsyncMock(return_value=4842587.791645115)
+    async def side_effect_function(*args, **kwargs):
+        if args[1] == "EUBUCCO":
+            return 6000000.791645115
+        else:
+            return 5000000.791645115
+
+    async_mock = AsyncMock(side_effect=side_effect_function)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_comparison.indicator.get_eubucco_building_area",
+        "ohsome_quality_api.indicators.building_comparison.indicator.get_reference_building_area",
         side_effect=async_mock,
     )
 
 
-@pytest.fixture(scope="class")
+@pytest.fixture
 def mock_get_building_area_low(class_mocker):
-    async_mock = AsyncMock(return_value=1)
+    async_mock = AsyncMock(return_value=1000000)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_comparison.indicator.get_eubucco_building_area",
+        "ohsome_quality_api.indicators.building_comparison.indicator.get_reference_building_area",
         side_effect=async_mock,
     )
 
 
-@pytest.fixture(scope="class")
+@pytest.fixture
+def mock_get_building_area_low_some(class_mocker):
+    async def side_effect_function(*args, **kwargs):
+        if args[1] == "EUBUCCO":
+            return 1
+        else:
+            return 6000000.791645115
+
+    async_mock = AsyncMock(side_effect=side_effect_function)
+    class_mocker.patch(
+        "ohsome_quality_api.indicators.building_comparison.indicator.get_reference_building_area",
+        side_effect=async_mock,
+    )
+
+
+@pytest.fixture
 def mock_get_building_area_empty(class_mocker):
     async_mock = AsyncMock(return_value=0)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_comparison.indicator.get_eubucco_building_area",
+        "ohsome_quality_api.indicators.building_comparison.indicator.get_reference_building_area",
         side_effect=async_mock,
     )
 
 
-@pytest.fixture(scope="class")
-def mock_get_eubucco_coverage_intersection(class_mocker, feature_germany_berlin):
+@pytest.fixture
+def mock_get_intersection_geom(class_mocker, feature_germany_berlin):
     async_mock = AsyncMock(return_value=feature_germany_berlin)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_comparison.indicator.db_client.get_eubucco_coverage_intersection",
+        "ohsome_quality_api.indicators.building_comparison.indicator.db_client.get_intersection_geom",
         side_effect=async_mock,
     )
 
 
-@pytest.fixture(scope="class")
-def mock_get_eubucco_coverage_intersection_area(class_mocker):
-    async_mock = AsyncMock(return_value=[{"area_ratio": 1}])
+@pytest.fixture
+def mock_get_intersection_area(class_mocker):
+    async_mock = AsyncMock(return_value=1.0)
     class_mocker.patch(
-        "ohsome_quality_api.indicators.building_comparison.indicator.db_client.get_eubucco_coverage_intersection_area",
+        "ohsome_quality_api.indicators.building_comparison.indicator.db_client.get_intersection_area",
+        side_effect=async_mock,
+    )
+
+
+@pytest.fixture
+def mock_get_intersection_area_none(class_mocker):
+    async_mock = AsyncMock(return_value=0)
+    class_mocker.patch(
+        "ohsome_quality_api.indicators.building_comparison.indicator.db_client.get_intersection_area",
+        side_effect=async_mock,
+    )
+
+
+@pytest.fixture
+def mock_get_intersection_area_some(class_mocker):
+    async def side_effect(*args, **kwargs):
+        if "eubucco" in args[1]:
+            return 0.0  # 0 %
+        else:
+            return 1.0  # 100 %
+
+    async_mock = AsyncMock(side_effect=side_effect)
+    class_mocker.patch(
+        "ohsome_quality_api.indicators.building_comparison.indicator.db_client.get_intersection_area",
         side_effect=async_mock,
     )
 
@@ -64,39 +108,83 @@ class TestInit:
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         assert indicator.th_high == 0.85
         assert indicator.th_low == 0.5
+        assert isinstance(indicator.data_ref, dict)
+
+    def test_get_sources(self, topic_building_area, feature_germany_berlin):
+        indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
+        source = indicator.format_sources()
+        assert "<a href='https://docs.eubucco.com/'>EUBUCCO</a>" in source
+
+    def test_attribution(self, topic_building_area, feature_germany_berlin):
+        indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
+        assert indicator.attribution is not None
+        assert indicator.attribution != ""
 
 
 class TestPreprocess:
     @oqapi_vcr.use_cassette
-    def test_preprocess(
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_area",
+        "mock_get_intersection_geom",
+    )
+    def test_preprocess(self, topic_building_area, feature_germany_berlin):
+        indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
+        asyncio.run(indicator.preprocess())
+
+        for area in indicator.area_osm.values():
+            assert area is not None
+            assert area > 0
+        assert isinstance(indicator.result.timestamp, datetime)
+        assert isinstance(indicator.result.timestamp_osm, datetime)
+
+    @oqapi_vcr.use_cassette
+    def test_preprocess_no_intersection(
         self,
         mock_get_building_area,
         topic_building_area,
         feature_germany_berlin,
-        mock_get_eubucco_coverage_intersection_area,
-        mock_get_eubucco_coverage_intersection,
+        mock_get_intersection_area_none,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         asyncio.run(indicator.preprocess())
 
-        assert indicator.area_osm is not None
-        assert indicator.area_osm > 0
-        assert indicator.area_references == {
-            "EUBUCCO": 4.842587791645116,
-        }
+        for area in indicator.area_cov.values():
+            assert area == 0.0
+        assert isinstance(indicator.result.timestamp, datetime)
+        assert indicator.result.timestamp_osm is None
+
+    @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_geom",
+        "mock_get_intersection_area_some",
+    )
+    def test_preprocess_some_intersection(
+        self,
+        topic_building_area,
+        feature_germany_berlin,
+    ):
+        indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
+        asyncio.run(indicator.preprocess())
+
+        assert 1.0 in list(indicator.area_cov.values())
+        assert 0.0 in list(indicator.area_cov.values())
         assert isinstance(indicator.result.timestamp, datetime)
         assert isinstance(indicator.result.timestamp_osm, datetime)
 
 
 class TestCalculate:
     @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_area",
+        "mock_get_intersection_geom",
+    )
     def test_calculate(
         self,
-        mock_get_building_area,
         topic_building_area,
         feature_germany_berlin,
-        mock_get_eubucco_coverage_intersection_area,
-        mock_get_eubucco_coverage_intersection,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         asyncio.run(indicator.preprocess())
@@ -104,82 +192,172 @@ class TestCalculate:
         assert indicator.result.value is not None
         assert indicator.result.value > 0
         assert indicator.result.class_ is not None
-        assert indicator.result.class_ == 5
+        assert indicator.result.class_ >= 0
 
     @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures(
+        "mock_get_building_area_empty",
+        "mock_get_intersection_area",
+        "mock_get_intersection_geom",
+    )
     def test_calculate_reference_area_0(
         self,
-        mock_get_building_area_empty,
         topic_building_area,
         feature_germany_heidelberg,
-        mock_get_eubucco_coverage_intersection_area,
-        mock_get_eubucco_coverage_intersection,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_heidelberg)
         asyncio.run(indicator.preprocess())
-        indicator.coverage["EUBUCCO"] is None
         indicator.calculate()
-        assert indicator.result.value is None
+        assert indicator.result.value == 0.0
 
     @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures(
+        "mock_get_intersection_area",
+        "mock_get_intersection_geom",
+        "mock_get_building_area_low",
+    )
     def test_calculate_above_one_th(
         self,
-        mock_get_building_area_low,
         topic_building_area,
         feature_germany_heidelberg,
-        mock_get_eubucco_coverage_intersection_area,
-        mock_get_eubucco_coverage_intersection,
+    ):
+        indicator = BuildingComparison(topic_building_area, feature_germany_heidelberg)
+        asyncio.run(indicator.preprocess())
+        indicator.calculate()
+        assert indicator.result.value is None
+        assert indicator.result.class_ is None
+        assert indicator.result.description is not None
+        for v in indicator.area_ref.values():
+            assert v is not None
+        for v in indicator.area_osm.values():
+            assert v is not None
+        assert (
+            "Warning: OSM has substantivly more buildings mapped than the Reference "
+            "datasets. No quality estimation has been made."
+            in indicator.result.description
+        )
+        assert indicator.result.label == "undefined"
+
+    @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures("mock_get_intersection_area_none")
+    def test_calculate_no_intersection(
+        self,
+        topic_building_area,
+        feature_germany_heidelberg,
+    ):
+        indicator = BuildingComparison(topic_building_area, feature_germany_heidelberg)
+        asyncio.run(indicator.preprocess())
+        indicator.calculate()
+        assert indicator.result.value is None
+        assert indicator.result.class_ is None
+        assert indicator.result.description is not None
+        assert (
+            "Comparison could not be made. None of the reference datasets covers the "
+            "area-of-interest." in indicator.result.description
+        )
+        assert indicator.result.label == "undefined"
+
+    @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_geom",
+        "mock_get_intersection_area_some",
+    )
+    def test_calculate_some_intersection(
+        self,
+        topic_building_area,
+        feature_germany_heidelberg,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_heidelberg)
         asyncio.run(indicator.preprocess())
         indicator.calculate()
         assert indicator.result.value is not None
         assert indicator.result.value > 0
-        assert indicator.result.class_ is None
+        assert indicator.result.class_ is not None
+        assert indicator.result.class_ >= 0
         assert indicator.result.description is not None
+        # major edge case description
         assert (
-            "Warning: Because of a big difference between OSM and the reference "
-            + "data no quality estimation has been made. "
-            + "It could be that the reference data is outdated. "
+            "Reference dataset EUBUCCO does not cover area-of-interest."
             in indicator.result.description
         )
-        assert indicator.result.label == "undefined"
+        assert (
+            "of the area-of-interest is covered by the reference dataset"
+            in indicator.result.description
+        )
+
+    @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_geom",
+        "mock_get_intersection_area",
+        "mock_get_building_area_low_some",
+    )
+    def test_calculate_above_one_th_and_expected(
+        self,
+        topic_building_area,
+        feature_germany_heidelberg,
+    ):
+        indicator = BuildingComparison(topic_building_area, feature_germany_heidelberg)
+        asyncio.run(indicator.preprocess())
+        indicator.calculate()
+        assert indicator.result.value is not None
+        assert indicator.result.value > 0
+        assert indicator.result.class_ is not None
+        assert indicator.result.class_ >= 0
+        assert indicator.result.label != "undefined"
+        assert indicator.result.description is not None
+        assert (
+            "of the area-of-interest is covered by the reference dataset"
+            in indicator.result.description
+        )
+        # TODO: should user be warned if a dataset was not fit for comparison?
+        # -> for ratio above self.above_one_th
+        # assert (
+        #     "this data is not considered in the overall result value."
+        #     in indicator.result.description
+        # )
 
 
 class TestFigure:
-    @pytest.fixture(scope="class")
     @oqapi_vcr.use_cassette
-    def indicator(
-        self,
-        mock_get_building_area,
-        topic_building_area,
-        feature_germany_berlin,
-        mock_get_eubucco_coverage_intersection_area,
-        mock_get_eubucco_coverage_intersection,
-    ):
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_geom",
+        "mock_get_intersection_area",
+    )
+    def test_create_figure(self, topic_building_area, feature_germany_berlin):
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         asyncio.run(indicator.preprocess())
         indicator.calculate()
-        return indicator
-
-    def test_create_figure(self, indicator):
         indicator.create_figure()
         assert isinstance(indicator.result.figure, dict)
         pgo.Figure(indicator.result.figure)  # test for valid Plotly figure
 
+    @oqapi_vcr.use_cassette
     @pytest.mark.skip(reason="Only for manual testing.")  # comment for manual test
-    def test_create_figure_manual(self, indicator):
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_geom",
+        "mock_get_intersection_area",
+    )
+    def test_create_figure_manual(self, topic_building_area, feature_germany_berlin):
+        indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
+        asyncio.run(indicator.preprocess())
+        indicator.calculate()
         indicator.create_figure()
         pio.show(indicator.result.figure)
 
     @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_geom",
+        "mock_get_intersection_area",
+    )
     def test_create_figure_above_one_th(
         self,
-        mock_get_building_area_low,
         topic_building_area,
         feature_germany_berlin,
-        mock_get_eubucco_coverage_intersection_area,
-        mock_get_eubucco_coverage_intersection,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         asyncio.run(indicator.preprocess())
@@ -190,13 +368,15 @@ class TestFigure:
         pgo.Figure(indicator.result.figure)
 
     @oqapi_vcr.use_cassette
+    @pytest.mark.usefixtures(
+        "mock_get_building_area",
+        "mock_get_intersection_geom",
+        "mock_get_intersection_area",
+    )
     def test_create_figure_building_area_zero(
         self,
-        mock_get_building_area_empty,
         topic_building_area,
         feature_germany_berlin,
-        mock_get_eubucco_coverage_intersection_area,
-        mock_get_eubucco_coverage_intersection,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         asyncio.run(indicator.preprocess())
@@ -205,8 +385,3 @@ class TestFigure:
         assert isinstance(indicator.result.figure, dict)
         assert indicator.result.figure["data"][0]["type"] == "bar"
         pgo.Figure(indicator.result.figure)
-
-
-def test_get_sources():
-    source = get_sources(["EUBUCCO"])
-    assert source == "<a href='https://docs.eubucco.com'>EUBUCCO</a>"